[Feature][Branch][Slider] Split [Rail] to [Rail] and [Marks] (#125)
* Split [Rail] to [Rail] and [Marks].

* Rename Rail test to Marks.

* Update snapshot test for [Slider].
