+++ src/components/shared/Slider/Marks.tsx
@@ -1,0 +1,354 @@
+import React, { FC, useEffect, useMemo } from 'react';
+import { StyleProp, StyleSheet, TouchableWithoutFeedback, ViewStyle } from 'react-native';
+
+import styled from 'styled-components/native';
+
+interface MarkPositionerType {
+  width: number;
+  position: number;
+}
+
+interface Props {
+  testID?: string;
+  sliderWidth: number;
+  style?: StyleProp<ViewStyle>;
+  mark?: React.ReactElement;
+  customMarkWidth?: number;
+  step?: number;
+  pixelsPerStep?: number;
+  markCount?: number;
+  startMark?: boolean;
+  endMark?: boolean;
+  disabled?: boolean;
+  onInit?: (markValues: number[], markPositions: number[]) => void | Promise<void>;
+  onMarkPress?: (value: number, position: number, index: number) => void | Promise<void>;
+}
+
+const DEFAULT = {
+  width: 3,
+  height: 3,
+};
+
+const Container = styled.View`
+  display: flex;
+  justify-content: center;
+  position: absolute;
+`;
+
+const MarkPositioner = styled.View<MarkPositionerType>`
+  position: absolute;
+  display: flex;
+  align-items: center;
+  width: ${({ width }): number => width};
+  left: ${({ position }): number => position};
+`;
+
+const Mark = styled.View`
+  width: ${DEFAULT.width};
+  height: ${DEFAULT.height};
+  background-color: #4163F4;
+`;
+
+// Checks if null or undefined. temporary until using othe library like 'lodash'.
+const isNil = (value: any): boolean => value == null;
+
+const getMarkCountByStep = ({
+  railWidth,
+  step,
+  startMark,
+  endMark,
+}: {
+  railWidth: number;
+  step: number;
+  startMark: boolean;
+  endMark: boolean;
+}): number => {
+  const count = Math.floor(railWidth / step);
+  const countToApply = (count <= 0) ? 2 : count + 1;
+
+  return countToApply - (startMark ? 0 : 1) - (endMark ? 0 : 1);
+};
+
+const getStepByMarkCount = ({
+  railWidth,
+  count,
+  startMark,
+  endMark,
+}): number => {
+  const step = railWidth / count;
+
+  if (startMark && endMark) {
+    return railWidth / (count - 1);
+  }
+
+  if (!startMark && !endMark) {
+    return (railWidth - step) / count;
+  }
+
+  return (railWidth - step) / (count - 1);
+};
+
+const getStepDistanceByMarkCount = ({
+  railWidth,
+  markWidth,
+  startMark,
+  endMark,
+  count,
+}: {
+  railWidth: number;
+  markWidth: number;
+  startMark: boolean;
+  endMark: boolean;
+  count: number;
+}): number => {
+  const getStepFineTunedStepByMark = ({
+    markWidth,
+    markCount,
+    step,
+  }: {
+    markWidth: number;
+    markCount: number;
+    step: number;
+  }): number => {
+    const fineTunerToFitMark = markWidth / markCount;
+
+    return step - fineTunerToFitMark;
+  };
+
+  const stepToApply = getStepByMarkCount({
+    railWidth,
+    count,
+    startMark,
+    endMark,
+  });
+
+  return getStepFineTunedStepByMark({
+    markWidth,
+    markCount: count,
+    step: stepToApply,
+  });
+};
+
+const getStepDistanceByStep = ({
+  railWidth,
+  markWidth,
+  markCount,
+  startMark,
+  endMark,
+}: {
+  railWidth: number;
+  markWidth: number;
+  markCount: number;
+  startMark: boolean;
+  endMark: boolean;
+}): number => {
+  return getStepDistanceByMarkCount({
+    railWidth,
+    markWidth,
+    startMark,
+    endMark,
+    count: markCount,
+  });
+};
+
+const getStepDistance = ({
+  railWidth,
+  markWidth,
+  step,
+  markCount,
+  startMark,
+  endMark,
+}: {
+  railWidth: number;
+  markWidth: number;
+  step: number;
+  markCount?: number;
+  startMark: boolean;
+  endMark: boolean;
+}): {
+  markCount: number;
+  stepDistance: number;
+} => {
+  const options = {
+    railWidth,
+    markWidth,
+    startMark,
+    endMark,
+  };
+
+  if (isNil(markCount)) {
+    const count = getMarkCountByStep({
+      railWidth,
+      step,
+      startMark,
+      endMark,
+    });
+
+    return {
+      markCount: count,
+      stepDistance: getStepDistanceByStep({
+        ...options,
+        markCount: count,
+      }),
+    };
+  };
+
+  return {
+    markCount: markCount as number,
+    stepDistance: getStepDistanceByMarkCount({
+      ...options,
+      count: markCount as number,
+    }),
+  };
+};
+
+const getMarkPositions = ({
+  startMark,
+  count,
+  stepDistance,
+}: {
+  startMark: boolean;
+  count: number;
+  stepDistance: number;
+}): number[] => {
+  const startAt = startMark ? 0 : stepDistance;
+
+  return Array.from({ length: count }).map((_, index: number) => startAt + (stepDistance * index));
+};
+
+const getMarkValue = (step: number, markIndex: number): number => step * markIndex;
+const getMarkValues = (
+  step: number,
+  markPositions: number[],
+): number[] => markPositions.map((_, index) => getMarkValue(step, index));
+
+const createMarks = ({
+  positions,
+  step,
+  stepDistance,
+  style,
+  mark,
+  markWidth,
+  disabled,
+  onMarkPress,
+}: {
+  positions: number[];
+  step: number;
+  stepDistance: number;
+  style: ViewStyle;
+  mark?: React.ReactElement;
+  markWidth: number;
+  disabled?: boolean;
+  onMarkPress?: (value: number, position: number, index: number) => void | Promise<void>;
+}): React.ReactElement[] => {
+  const halfStepDistance = stepDistance / 2;
+  const fineTunedHalfStepDistance = halfStepDistance - (markWidth / 2);
+
+  return positions.map((position: number, index: number): React.ReactElement => {
+    const handlePress = (): void => {
+      if (disabled) {
+        return;
+      }
+
+      const value = getMarkValue(step, index);
+      if (onMarkPress) {
+        onMarkPress(value, position, index);
+      }
+    };
+
+    return (
+      <TouchableWithoutFeedback key={position} onPress={handlePress}>
+        <MarkPositioner width={stepDistance} position={position - fineTunedHalfStepDistance}>
+          {mark || <Mark style={style} />}
+        </MarkPositioner>
+      </TouchableWithoutFeedback>
+    );
+  });
+};
+
+const getMarkWidth = (markStyle: ViewStyle): number => {
+  const { width } = markStyle;
+
+  if (isNil(width)) {
+    return DEFAULT.width;
+  }
+
+  return parseInt((width as string | number).toString());
+};
+
+const Marks: FC<Props> = ({
+  testID,
+  sliderWidth,
+  style = {},
+  mark,
+  customMarkWidth,
+  step = 20,
+  pixelsPerStep,
+  markCount,
+  startMark = true,
+  endMark = true,
+  disabled = false,
+  onInit,
+  onMarkPress,
+}) => {
+  const handleInit = (markValues: number[], markPositions: number[]): void => {
+    if (onInit) {
+      onInit(markValues, markPositions);
+    }
+  };
+
+  if (mark && isNil(customMarkWidth)) {
+    throw Error('Must have customMarkWidth prop if it uses a cutsom mark.');
+  }
+
+  const markStyleToApply = StyleSheet.flatten(style);
+
+  const railWidth = sliderWidth;
+  const markWidth = isNil(mark) ? getMarkWidth(markStyleToApply) : customMarkWidth as number;
+
+  const stepByPixel = isNil(pixelsPerStep) ? step : step * (pixelsPerStep as number);
+
+  const markOptions = {
+    railWidth,
+    markWidth,
+    step: stepByPixel,
+    startMark,
+    endMark,
+  };
+
+  const { markCount: markCountToApply, stepDistance } = useMemo(() => getStepDistance({
+    ...markOptions,
+    markCount,
+  }), Object.values(markOptions));
+
+  const markPositions = useMemo(() => getMarkPositions({
+    startMark,
+    count: markCountToApply,
+    stepDistance,
+  }), [startMark, markCountToApply, stepDistance]);
+
+  const marks = createMarks({
+    positions: markPositions,
+    step,
+    stepDistance,
+    style: markStyleToApply,
+    mark,
+    markWidth,
+    disabled,
+    onMarkPress,
+  });
+
+  useEffect(() => {
+    const markValues = getMarkValues(step, markPositions);
+
+    handleInit(markValues, markPositions);
+  }, []);
+
+  return (
+    <Container testID={testID}>
+      {marks}
+    </Container>
+  );
+};
+
+export default Marks;
